230610: decorate input element to open folder
